+++ src/main/java/com/jaquadro/minecraft/storagedrawers/api/security/ISecurityProvider.java
@@ -2,6 +2,7 @@ package com.jaquadro.minecraft.storagedrawers.api.security;
 
 import com.jaquadro.minecraft.storagedrawers.api.storage.attribute.IProtectable;
 import com.mojang.authlib.GameProfile;
+import net.minecraft.world.entity.player.Player;
 
 public interface ISecurityProvider
 {
@@ -9,5 +10,5 @@ public interface ISecurityProvider
 
     boolean hasOwnership (GameProfile profile, IProtectable target);
 
-    boolean hasAccess (GameProfile profile, IProtectable target);
+    boolean hasAccess (Player player, IProtectable target);
 }