+++ src/main/java/com/jaquadro/minecraft/storagedrawers/item/ItemPersonalKey.java
@@ -1,59 +1,41 @@
-/*package com.jaquadro.minecraft.storagedrawers.item;
+package com.jaquadro.minecraft.storagedrawers.item;
 
-import com.google.common.collect.Multimap;
-import com.jaquadro.minecraft.storagedrawers.core.ModCreativeTabs;
-import net.minecraft.client.resources.I18n;
-import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.entity.SharedMonsterAttributes;
-import net.minecraft.entity.ai.attributes.AttributeModifier;
-import net.minecraft.inventory.EntityEquipmentSlot;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.world.World;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
+import com.jaquadro.minecraft.storagedrawers.api.storage.IDrawerAttributesModifiable;
+import com.jaquadro.minecraft.storagedrawers.integration.LocalIntegrationRegistry;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.item.Item;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+import org.jetbrains.annotations.NotNull;
 
-import javax.annotation.Nonnull;
-import org.jetbrains.annotations.Nullable;
-import java.util.List;
-
-public class ItemPersonalKey extends Item
+public class ItemPersonalKey extends ItemKey
 {
-    public ItemPersonalKey (String registryName, String unlocalizedName) {
-        setRegistryName(registryName);
-        setUnlocalizedName(unlocalizedName);
-        setCreativeTab(ModCreativeTabs.tabStorageDrawers);
-        setMaxDamage(0);
+    private final String securityProvider;
+
+    public ItemPersonalKey (String securityProvider, Item.Properties properties) {
+        super(properties);
+        this.securityProvider = securityProvider;
     }
 
     @Override
-    @SideOnly(Side.CLIENT)
-    public void addInformation (@Nonnull ItemStack itemStack, @Nullable World world, List<String> list, ITooltipFlag advanced) {
-        String name = getUnlocalizedName(itemStack);
-        list.add(I18n.format(name + ".description"));
+    protected void handleDrawerAttributes (IDrawerAttributesModifiable attrs) {
+        attrs.setIsShowingQuantity(!attrs.isShowingQuantity());
     }
 
-    @SideOnly(Side.CLIENT)
-    public boolean isFull3D()
-    {
-        return true;
+    public String getSecurityProviderKey () {
+        return securityProvider;
     }
 
     @Override
-    public Multimap<String, AttributeModifier> getAttributeModifiers (EntityEquipmentSlot slot, @Nonnull ItemStack stack) {
-        Multimap<String, AttributeModifier> multimap = super.getAttributeModifiers(slot, stack);
-
-        if (slot == EntityEquipmentSlot.MAINHAND)
-            multimap.put(SharedMonsterAttributes.ATTACK_DAMAGE.getName(), new AttributeModifier(ATTACK_DAMAGE_MODIFIER, "Weapon modifier", (double)2, 0));
-
-        return multimap;
-    }
-
-    public String getSecurityProviderKey (int meta) {
-        switch (meta) {
-            case 0: return null;
-            default: return null;
+    @OnlyIn(Dist.CLIENT)
+    @NotNull
+    public Component getDescription() {
+        if (securityProvider != null) {
+            if (securityProvider.equals("cofh") && !LocalIntegrationRegistry.isModLoaded("cofh_core"))
+                return Component.translatable("itemConfig.storagedrawers.disabled_upgrade").withStyle(ChatFormatting.RED);
         }
+
+        return super.getDescription();
     }
 }
-*/